Updates Elevator Subsystem
This commit updates the elevator subsystem with fixes and changes from testing.
It fixes an issue where PID was done on the end set point rather than expected position based on the motion profile.
It explicitly sets the idle mode, current limits, and enables limit switches for the elevator.
It removes the DIO limit switches since they are now fed through the SparkFlex.